6.0 Configurable Beacon Radius (#10)
* move from multiplier system on xp drain to xp points per amplitude per beacon tick

* No more 0 amplifier

* Add configurable beacon radius according to the 4 beacon levels

* README update

+++ src/main/java/net/fracturedcode/xpbeacons/EffectSettings.java
@@ -11,6 +11,6 @@ public abstract class EffectSettings {
     public abstract int getEffectAmplitudeCeiling();
     public abstract int getEffectXpCeiling();
     public abstract boolean getXpAmplitudeToggle();
-    public abstract double getXpDrainRate();
+    public abstract int getXpDrainRate();
     public abstract boolean getShouldDrainXp();
 }
